# packages/shared/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MetricSummaryOut": {
        "properties": {
          "definition": {
            "$ref": "#/components/schemas/MetricDefinitionOut"
          },
          "latest": {
            "$ref": "#/components/schemas/MetricValue"
          },
          "record": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonalRecordOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "window": {
            "$ref": "#/components/schemas/MetricComparison"
          }
        },
        "required": [
          "definition",
          "latest",
          "window",
          "record"
        ],
        "title": "MetricSummaryOut",
        "type": "object"
      },
      "MetricValue": {
        "description": "A number that always travels with its unit and its evidence.",
        "properties": {
          "observed_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed On"
          },
          "sample_size": {
            "default": 0,
            "title": "Sample Size",
            "type": "integer"
          },
          "source_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "value",
          "unit"
        ],
        "title": "MetricValue",
        "type": "object"
      },
      "PersonalRecordEventOut": {
        "properties": {
          "achieved_on": {
            "format": "date",
            "title": "Achieved On",
            "type": "string"
          },
          "delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta"
          },
          "display_precision": {
            "title": "Display Precision",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "metric_display_name": {
            "title": "Metric Display Name",
            "type": "string"
          },
          "metric_key": {
            "title": "Metric Key",
            "type": "string"
          },
          "new_value": {
            "title": "New Value",
            "type": "number"
          },
          "player_display_name": {
            "title": "Player Display Name",
            "type": "string"
          },
          "player_id": {
            "format": "uuid",
            "title": "Player Id",
            "type": "string"
          },
          "previous_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Value"
          },
          "sample_size": {
            "title": "Sample Size",
            "type": "integer"
          },
          "source_status": {
            "$ref": "#/components/schemas/SourceStatus"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "id",
          "player_id",
          "player_display_name",
          "metric_key",
          "metric_display_name",
          "unit",
          "display_precision",
          "previous_value",
          "new_value",
          "delta",
          "sample_size",
          "achieved_on",
          "source_status"
        ],
        "title": "PersonalRecordEventOut",
        "type": "object"
      },
      "PersonalRecordOut": {
        "properties": {
          "achieved_on": {
            "format": "date",
            "title": "Achieved On",
            "type": "string"
          },
          "display_precision": {
            "title": "Display Precision",
            "type": "integer"
          },
          "is_manual_override": {
            "default": false,
            "title": "Is Manual Override",
            "type": "boolean"
          },
          "metric_display_name": {
            "title": "Metric Display Name",
            "type": "string"
          },
          "metric_key": {
            "title": "Metric Key",
            "type": "string"
          },
          "sample_size": {
            "title": "Sample Size",
            "type": "integer"
          },
          "session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "source_status": {
            "$ref": "#/components/schemas/SourceStatus"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "metric_key",
          "metric_display_name",
          "unit",
          "display_precision",
          "value",
          "sample_size",
          "achieved_on",
          "session_id",
          "source_status"
        ],
        "title": "PersonalRecordOut",
        "type": "object"
      },
      "PlayerOverviewOut": {
        "properties": {
          "last_session": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricSummaryOut"
            },
            "title": "Metrics",
            "type": "array"
          },
          "player": {
            "$ref": "#/components/schemas/PlayerSummary"
          },
          "sessions_in_window": {
            "title": "Sessions In Window",
            "type": "integer"
          },
          "time_range": {
            "title": "Time Range",
            "type": "string"
          }
        },
        "required": [
          "player",
          "time_range",
          "last_session",
          "sessions_in_window",
          "metrics"
        ],
        "title": "PlayerOverviewOut",
        "type": "object"
      },
      "PlayerSummary": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "bats": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bats"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "graduation_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graduation Year"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "preferred_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Name"
          },
          "throws": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Throws"
          }
        },
        "required": [
          "id",
          "display_name",
          "full_name",
          "first_name",
          "last_name",
          "active"
        ],
        "title": "PlayerSummary",
        "type": "object"
      },
...
      "SessionOut": {
        "properties": {
          "hit_count": {
            "default": 0,
            "title": "Hit Count",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "pitch_count": {
            "default": 0,
            "title": "Pitch Count",
            "type": "integer"
          },
          "session_date": {
            "format": "date",
            "title": "Session Date",
            "type": "string"
          },
          "session_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Type"
          },
          "source_status": {
            "$ref": "#/components/schemas/SourceStatus"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "venue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue"
          },
          "verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          }
        },
        "required": [
          "id",
          "session_date",
          "source_status"
        ],
        "title": "SessionOut",
        "type": "object"
      },
      "SourceStatus": {
        "description": "Confidence level of vendor-supplied data.\n\nTrackMan can publish a session as PRELIMINARY and later republish it as\nVERIFIED with corrected values. Everything derived from an event carries the\nstatus of the data it was derived from so the UI can label it honestly.",
        "enum": [
          "PRELIMINARY",
          "VERIFIED"
        ],
        "title": "SourceStatus",
        "type": "string"
      },
...
      "TimeRange": {
        "description": "Windows the dashboards offer.",
        "enum": [
          "today",
          "7d",
          "30d",
          "90d",
          "1y",
          "all"
        ],
        "title": "TimeRange",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Internal player analytics for a baseball training facility. All athlete access is authorized server-side and scoped to one organization.",
    "title": "Baseball Analytics API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me/overview": {
      "get": {
        "operationId": "my_overview_api_v1_me_overview_get",
        "parameters": [
          {
            "in": "query",
            "name": "range",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TimeRange",
              "default": "30d"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerOverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "My Overview",
        "tags": [
          "me"
        ]
      }
    },
...
    "/api/v1/me/sessions": {
      "get": {
        "operationId": "my_sessions_api_v1_me_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionOut"
                  },
                  "title": "Response My Sessions Api V1 Me Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "My Sessions",
        "tags": [
          "me"
        ]
      }
    },
...
    "/api/v1/players": {
      "get": {
        "description": "Search the roster. Staff only -- a player has no roster.",
        "operationId": "list_players_api_v1_players_get",
        "parameters": [
          {
            "description": "name search: first, last or preferred",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "name search: first, last or preferred",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "active_only",
            "required": false,
            "schema": {
              "default": true,
              "title": "Active Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlayerSummary"
                  },
                  "title": "Response List Players Api V1 Players Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Players",
        "tags": [
          "players"
        ]
      }
    },
    "/api/v1/players/{player_id}": {
      "get": {
        "operationId": "get_player_api_v1_players__player_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Player Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Player",
        "tags": [
          "players"
        ]
      }
    },
...
    "/api/v1/players/{player_id}/overview": {
      "get": {
        "description": "Headline metrics, their movement over the window, and current records.",
        "operationId": "get_overview_api_v1_players__player_id__overview_get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Player Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "range",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TimeRange",
              "default": "30d"
            }
          },
          {
            "in": "query",
            "name": "headline_only",
            "required": false,
            "schema": {
              "default": true,
              "title": "Headline Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerOverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Overview",
        "tags": [
          "players"
        ]
      }
    },
    "/api/v1/players/{player_id}/prs": {
      "get": {
        "operationId": "get_personal_records_api_v1_players__player_id__prs_get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Player Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersonalRecordOut"
                  },
                  "title": "Response Get Personal Records Api V1 Players  Player Id  Prs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Personal Records",
        "tags": [
          "players"
        ]
      }
    },
    "/api/v1/players/{player_id}/prs/{metric_key}/history": {
      "get": {
        "description": "The full progression -- when each record was set, and by how much.",
        "operationId": "get_record_history_api_v1_players__player_id__prs__metric_key__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Player Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "metric_key",
            "required": true,
            "schema": {
              "title": "Metric Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersonalRecordEventOut"
                  },
                  "title": "Response Get Record History Api V1 Players  Player Id  Prs  Metric Key  History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Record History",
        "tags": [
          "players"
        ]
      }
    },
    "/api/v1/players/{player_id}/sessions": {
      "get": {
        "operationId": "get_sessions_api_v1_players__player_id__sessions_get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Player Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionOut"
                  },
                  "title": "Response Get Sessions Api V1 Players  Player Id  Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Sessions",
        "tags": [
          "players"
        ]
      }
    },
...
    "/api/v1/sessions": {
      "get": {
        "operationId": "list_sessions_api_v1_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionOut"
                  },
                  "title": "Response List Sessions Api V1 Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sessions",
        "tags": [
          "dashboard"
        ]
      }
    },
...
  }
}